Remove excessive laziness (#69)
Partially reverts
a2a3a0c7e47cf51da402aee7483071ed35d11f0d

Go back to unpacking scheduled queues in catenable queues.

I mistakenly thought that the scheduled queues had to be suspended
lazily in the tails of the catenable queues. This was incorrect. The
only changes that were actually required were to stop trying to discard
`Empty` in the second argument of `><` and to adjust `linkAll` to deal
with the possible presence of `Empty`.